Icônes Material Design (469) en Tiny TTF — parité designer↔firmware (#42)
* docs(icones): spec design — icônes MDI en Tiny TTF

Brainstorm validé : remplacer les 23 symboles LVGL bitmap par un catalogue
Material Design Icons (~300-500) rendu via Tiny TTF, pipeline généré depuis une
liste de noms (police firmware + webfont designer + mapping + métadonnées picker
+ enum schéma), parité designer↔firmware, picker recherche/catégories, enum
symbolName régénéré (validation stricte).


* docs(icones): plan d'implémentation — icônes MDI en Tiny TTF

Plan TDD task-par-task en 4 phases : A) pipeline gen_icons.py (sélection par
tags+extras+alias fail-loud, subset TTF, tables C, woff2, icons-data.js, enum
schéma) ; B) firmware (uint16_t, get_icon_font, rendu via glyphes générés) ;
C) designer (buildIcon webfont, picker recherche/catégories, parité, staging) ;
D) vérif native+node+browser+on-device. Points ouverts résolus en mécanismes
exécutables (version MDI épinglée, composition par tags, table d'alias legacy).


* feat(icones): gen_icons.py — pipeline MDI (source unique → tous artefacts)

Source unique tools/icons/{categories,extra,aliases}.txt → sélection =
(icônes portant un tag ∈ categories) ∪ extras ∪ cibles d'alias legacy,
dédupliquée par nom exposé, triée, plafonnée à MAX_ICONS=500. Fail-loud
si un extra/alias cible une icône absente de meta.json.

Set retenu : Weather + Battery + Alert/Error + Notification → 469 icônes
(les tags MDI étant quasi disjoints, l'union ≈ somme ; « Home Automation »
seul = 659 > cap, donc non inclus en bloc). 23 alias legacy préservent la
compat des layouts (wifi, gps, battery_1, volume_max, settings…).

Émet : police firmware (font_icons.c), tables C (icons_gen.{h,c}), webfont
designer (mdi.woff2), métadonnées picker (icons-data.js), enum $defs/symbolName
(remplacement ciblé format-préservant), licence MDI. MDI_VERSION=7.4.47.


* feat(icones): set MDI généré (469 icônes) — police firmware, webfont, tables, enum

Sorties committées de tools/gen_icons.py (régénérables, jamais éditées à la main) :
- src/fonts/font_icons.c (~65 Ko TTF subsetté) + décl. dans fonts_data.h
- src/fonts/icons_gen.{h,c} : ICON_SYMBOL_NAMES / ICON_GLYPHS (UTF-8) / ICON_SYMBOL_COUNT
- designer/vendor/fonts/mdi.woff2 (23 Ko, parité) + designer/vendor/icons/icons-data.js
- schema : $defs/symbolName régénéré (diff localisé au seul bloc)
- tools/fonts/licenses/mdi-LICENSE.txt

Parité stricte vérifiée : schéma == icons_gen.c == icons-data.js (469, même ordre).


* fix(icones): icons_gen.h C-safe (#define + extern "C")

icons_gen.c est compilé comme C (comme les font_*.c) ; le header utilisait
`static constexpr` (C++), qui échoue en C. Passage à `#define ICON_SYMBOL_COUNT`
+ garde extern "C" autour des externs, sur le modèle exact de fonts_data.h.
Inclus côté firmware/tests en C++ (link C). icons_gen.c inchangé.


* feat(icones): index icône en uint16_t + tables générées (firmware + tests)

Le set MDI (469) dépasse 255 : l'index d'icône passe de uint8_t à uint16_t sur
TOUTE la chaîne, sinon troncature silencieuse (ICON_GLYPHS[sym & 0xFF]).
- dashboard.{h,cpp} : IconState.symbol / Component.icon_symbol en uint16_t ;
  tables locales supprimées → icons_gen.h (généré) ; icon_symbol_index exposé (uint16_t).
- color.{h,cpp} : icon_resolve base_sym/out_sym/sym en uint16_t (non couvert par le plan,
  découvert en lisant les appelants — sinon tronque les états d'icône).
- platformio.ini : env native compile fonts/icons_gen.c (link des tables pour les tests).
- tests : icon_resolve/icon_parsed en uint16_t + assertions round-trip nom↔index
  (robustes au set) ; nouveau test_icon_symbol_index_legacy (compat des 23 noms).

189 tests natifs OK.


* feat(icones): rendu firmware via get_icon_font + glyphes MDI générés

- fonts.{h,cpp} : get_icon_font(px) — Tiny TTF sur font_icons, cache partagé avec
  get_font (famille=4 réservée), repli bitmap si échec.
- view.cpp : build_icon rend le glyphe via get_icon_font ; ICON_GLYPHS vient du header
  généré (glyphes MDI UTF-8). Supprime la table locale LV_SYMBOL_* et pick_font (mort
  après le switch, unique appelant).

Build esp32s3 OK (flash 61,6 %).


* feat(icones): rendu designer via webfont MDI + logique de filtre pure

- render.js : buildIcon rend le glyphe MDI via <i class="mdi"> (webfont), export ICON_CHAR
  (nom -> glyphe) ; supprime ICON_SVG + batterySvg (morts). resolveIcon inchangé.
- style.css : @font-face 'mdi' (vendor/fonts/mdi.woff2) + classe .mdi partagée
  (canvas, picker, inspecteur) + styles du picker/bouton-picker.
- icon-filter.js : filterIcons/categoriesOf (logique PURE, testée) + 5 tests node.


* feat(icones): picker d'icônes (recherche + catégories) dans l'inspecteur

- icon-picker.js : overlay MDI (recherche nom/tag + filtre catégorie + grille), allowClear
  pour l'option « base » des états. DOM transitoire (browser-verified), filtre pur testé.
- inspector.js : champ symbol (kind 'iconpicker') + table des états passent du <select> au
  bouton-picker (ref figée, mêmes invariants de commit). Retire ICON_SVG.
- registry.js : éditeur du champ symbol -> 'iconpicker'. i18n en/fr (iconpicker.*).
- registry.test.js : parité réécrite (enum symbolName == noms de icons-data.js).


* fix(icones): le picker ne désélectionne plus le composant au clic

L'overlay est monté sur document.body ; ses clics bubblaient jusqu'au handler
global de désélection du canvas (app.js : « pointerdown hors .w/#inspector/#layers
→ désélectionne »), qui vidait la sélection après chaque choix d'icône. Le picker
stoppe désormais la propagation du pointerdown (le clic-fond ferme toujours).
Vérifié au navigateur : après un pick, le composant reste sélectionné.

+++ src/color.h
@@ -8,5 +8,5 @@ uint32_t threshold_color(const Threshold* t, int n, float value, uint32_t base);
 bool led_is_lit(int32_t value, int32_t off_below);
 // icon : resout (glyphe, couleur) pour une valeur. 1re bande ou value < at ; champ omis -> base.
 // Miroir de threshold_color ; sans LVGL (testable en natif). symbol = index dans ICON_GLYPHS (view.cpp).
-void icon_resolve(const IconState* st, int n, float value, uint8_t base_sym, uint32_t base_col,
-                  uint8_t* out_sym, uint32_t* out_col);
+void icon_resolve(const IconState* st, int n, float value, uint16_t base_sym, uint32_t base_col,
+                  uint16_t* out_sym, uint32_t* out_col);
